ocr-demo/requirements.txt

google-cloud-vision